+++ lib/Storage/Exception/TagNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace PhpBench\Storage\Exception;
+
+use RuntimeException;
+
+class TagNotFoundException extends RuntimeException
+{
+}